+++ examples/widgets/lib/custom_list_view.dart
@@ -1,0 +1,101 @@
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'package:fl_utilities/fl_utilities.dart';
+
+void main() => runApp(const MyApp());
+
+class MyApp extends StatefulWidget {
+  const MyApp({super.key});
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  bool isVertDirection = true;
+  bool isReversed = false;
+  CustomListViewItemAlignment crossAxisAlignment =
+      CustomListViewItemAlignment.center;
+
+  @override
+  Widget build(BuildContext context) {
+    CrossAxisAlignment;
+
+    return MaterialApp(
+      theme: ThemeData(),
+      darkTheme: ThemeData.dark(),
+      themeMode: ThemeMode.system,
+
+      // allow mouse to drag.
+      scrollBehavior: context.scrollBehavior.copyWith(
+        dragDevices: {
+          ...context.scrollBehavior.dragDevices,
+          PointerDeviceKind.mouse,
+        },
+      ),
+      home: Scaffold(
+        appBar: AppBar(title: const Text('CustomListView Example')),
+        body: SizedBox.expand(
+          child: Column(
+            children: [
+              const SizedBox(height: 16.0),
+              SingleChildScrollView(
+                scrollDirection: Axis.horizontal,
+                child: Row(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    ElevatedButton(
+                      onPressed: () {
+                        setState(() => isVertDirection = !isVertDirection);
+                      },
+                      child: Text(
+                        isVertDirection
+                            ? 'scrollDirection: Axis.vertical'
+                            : 'scrollDirection: Axis.horizontal',
+                      ),
+                    ),
+                    const SizedBox(width: 16.0),
+                    ElevatedButton(
+                      onPressed: () => setState(() => isReversed = !isReversed),
+                      child: Text(
+                        isReversed ? 'reverse: true' : 'reverse: false',
+                      ),
+                    ),
+                    const SizedBox(width: 16.0),
+                    DropdownButton<CustomListViewItemAlignment>(
+                      value: crossAxisAlignment,
+                      onChanged: (value) => setState(() {
+                        crossAxisAlignment = value!;
+                      }),
+                      items: CustomListViewItemAlignment.values.map((e) {
+                        return DropdownMenuItem(
+                          value: e,
+                          child: Text('$e'),
+                        );
+                      }).toList(),
+                    )
+                  ],
+                ),
+              ),
+              const Divider(),
+              Expanded(
+                child: CustomListView.builder(
+                  scrollDirection:
+                      isVertDirection ? Axis.vertical : Axis.horizontal,
+                  reverse: isReversed,
+                  itemCount: 100,
+                  itemBuilder: (_, index) => CustomListViewItemDelegate(
+                    mainAxisLength: 200.0,
+                    crossAxisLength: 200.0,
+                    crossAxisAlignment: crossAxisAlignment,
+                    child: Card(child: Center(child: Text('#$index'))),
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}